container/requirements.txt
# PyPI library used in all project
rdetoolkit==1.5.1
matplotlib==3.10.1
plotly==6.0.1
defusedxml==0.7.1
requests==2.32.3
pillow==10.4.0
pydantic>=2.8.3
pandas==2.2.3
pandas-stubs
types-defusedxml
# PyPI libraries to be customized and installed in each project
# python-magic==0.4.25 ...etc
